+++ package.json
@@ -59,19 +59,19 @@
     "del": "^5.1.0",
     "del-cli": "^3.0.0",
     "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
+    "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-prettier": "^3.1.3",
     "husky": "^4.2.5",
     "inspect-loader": "^1.0.0",
     "jest": "^25.4.0",
-    "lint-staged": "^10.1.6",
+    "lint-staged": "^10.1.7",
     "memfs": "^3.1.2",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.4",
+    "prettier": "^2.0.5",
     "standard-version": "^7.1.0",
     "strip-ansi": "^6.0.0",
-    "webpack": "^4.42.1"
+    "webpack": "^4.43.0"
   },
   "keywords": [
     "webpack",
